backend/requirements.txt
django==4.2.14
djangorestframework==3.15.2
celery==5.4
redis==5.0.7
eventlet==0.36.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv
whitenoise